Move function to utils
